* cherry-picking merge from /branches/Performance to /trunk   r16738-16739,16741-16742,16767,16774,16776   (use AsycFramework in revision graph as an example    of how to speed up I/O and/or CPU-bound operations) * resolved a few conflicts
git-svn-id: https://svn.code.sf.net/p/tortoisesvn/code/trunk@16782 154be670-7969-47af-9ab0-c6fb378d2295

+++ src/TortoiseProc/Commands/CommitCommand.cpp
@@ -104,7 +104,7 @@ bool CommitCommand::Execute()
 			progDlg.SetCommitMessage(dlg.m_sLogMessage);
 			progDlg.SetDepth(dlg.m_bRecursive ? svn_depth_infinity : svn_depth_empty);
 			progDlg.SetSelectedList(dlg.m_selectedPathList);
-			progDlg.SetItemCount(dlg.m_itemsCount);
+			progDlg.SetItemCount(static_cast<long>(dlg.m_itemsCount));
 			progDlg.SetBugTraqProvider(dlg.m_BugTraqProvider);
 			progDlg.SetRevisionProperties(dlg.m_revProps);
 			progDlg.DoModal();
