+++ src/TortoiseProc/Settings/SetDialogs.cpp
@@ -202,7 +202,7 @@ BOOL CSetDialogs::OnApply()
 	else
 		m_sFontName = m_regFontName;
 
-    Store (m_dwAutoClose, m_regAutoClose);
+    Store (static_cast<DWORD>(m_dwAutoClose), m_regAutoClose);
 	Store (m_bShortDateFormat, m_regShortDateFormat);
     Store (m_bUseSystemLocaleForDates, m_regUseSystemLocaleForDates);
